fix: unified update transaction
