+++ src/pg/pg-store.ts
@@ -1,5 +1,6 @@
 import { Order, OrderBy } from '../api/schemas';
-import { SatoshiRarity } from '../api/util/ordinal-satoshi';
+import { OrdinalSatoshi, SatoshiRarity } from '../api/util/ordinal-satoshi';
+import { InscriptionEvent } from '../chainhook/schemas';
 import { ENV } from '../env';
 import { logger } from '../logger';
 import { getIndexResultCountType, inscriptionContentToJson } from './helpers';
@@ -48,16 +49,112 @@ export class PgStore extends BasePgStore {
     return new PgStore(sql);
   }
 
-  async updateChainTipBlockHeight(args: { blockHeight: number }): Promise<void> {
-    await this.sql`
-      UPDATE chain_tip SET block_height = GREATEST(${args.blockHeight}, block_height)
-    `;
-  }
-
-  async updateChainTipInscriptionCount(): Promise<void> {
-    await this.sql`
-      UPDATE chain_tip SET inscription_count = (SELECT COUNT(*) FROM inscriptions)
-    `;
+  /**
+   * Inserts inscription genesis and transfers from Chainhook events. Also handles rollbacks from
+   * chain re-orgs and materialized view refreshes.
+   * @param args - Apply/Rollback Chainhook events
+   */
+  async updateInscriptions(args: {
+    apply: InscriptionEvent[];
+    rollback: InscriptionEvent[];
+  }): Promise<void> {
+    const updatedInscriptionIds = new Set<number>();
+    await this.sqlWriteTransaction(async sql => {
+      for (const event of args.rollback) {
+        for (const tx of event.transactions) {
+          for (const operation of tx.metadata.ordinal_operations) {
+            if (operation.inscription_revealed) {
+              const genesis_id = operation.inscription_revealed.inscription_id;
+              await this.rollBackInscriptionGenesis({ genesis_id });
+              logger.info(`PgStore rollback inscription ${genesis_id}`);
+            }
+            if (operation.inscription_transferred) {
+              const genesis_id = operation.inscription_transferred.inscription_id;
+              const satpoint = operation.inscription_transferred.satpoint_post_transfer.split(':');
+              const output = `${satpoint[0]}:${satpoint[1]}`;
+              const id = await this.rollBackInscriptionTransfer({ genesis_id, output });
+              if (id) updatedInscriptionIds.add(id);
+              logger.info(`PgStore rollback transfer ${genesis_id} ${output}`);
+            }
+          }
+        }
+      }
+      for (const event of args.apply) {
+        for (const tx of event.transactions) {
+          for (const operation of tx.metadata.ordinal_operations) {
+            if (operation.inscription_revealed) {
+              const reveal = operation.inscription_revealed;
+              const txId = tx.transaction_identifier.hash.substring(2);
+              const satoshi = new OrdinalSatoshi(reveal.ordinal_number);
+              const id = await this.insertInscriptionGenesis({
+                inscription: {
+                  genesis_id: reveal.inscription_id,
+                  mime_type: reveal.content_type.split(';')[0],
+                  content_type: reveal.content_type,
+                  content_length: reveal.content_length,
+                  number: reveal.inscription_number,
+                  content: reveal.content_bytes,
+                  fee: reveal.inscription_fee.toString(),
+                },
+                location: {
+                  genesis_id: reveal.inscription_id,
+                  block_height: event.block_identifier.index,
+                  block_hash: event.block_identifier.hash.substring(2),
+                  tx_id: txId,
+                  address: reveal.inscriber_address,
+                  output: `${txId}:0`,
+                  offset: reveal.ordinal_offset.toString(),
+                  value: reveal.inscription_output_value.toString(),
+                  timestamp: event.timestamp,
+                  sat_ordinal: reveal.ordinal_number.toString(),
+                  sat_rarity: satoshi.rarity,
+                  sat_coinbase_height: satoshi.blockHeight,
+                },
+              });
+              if (id) updatedInscriptionIds.add(id);
+              logger.info(
+                `PgStore apply inscription #${reveal.inscription_number} (${reveal.inscription_id}) at block ${event.block_identifier.index}`
+              );
+            }
+            if (operation.inscription_transferred) {
+              const transfer = operation.inscription_transferred;
+              const txId = tx.transaction_identifier.hash.substring(2);
+              const satpoint = transfer.satpoint_post_transfer.split(':');
+              const offset = satpoint[2];
+              const output = `${satpoint[0]}:${satpoint[1]}`;
+              const satoshi = new OrdinalSatoshi(transfer.ordinal_number);
+              const id = await this.insertInscriptionTransfer({
+                location: {
+                  genesis_id: transfer.inscription_id,
+                  block_height: event.block_identifier.index,
+                  block_hash: event.block_identifier.hash,
+                  tx_id: txId,
+                  address: transfer.updated_address,
+                  output: output,
+                  offset: offset ?? null,
+                  value: transfer.post_transfer_output_value
+                    ? transfer.post_transfer_output_value.toString()
+                    : null,
+                  timestamp: event.timestamp,
+                  sat_ordinal: transfer.ordinal_number.toString(),
+                  sat_rarity: satoshi.rarity,
+                  sat_coinbase_height: satoshi.blockHeight,
+                },
+              });
+              if (id) updatedInscriptionIds.add(id);
+              logger.info(
+                `PgStore apply transfer for #${transfer.inscription_number} (${transfer.inscription_id}) to output ${output} at block ${event.block_identifier.index}`
+              );
+            }
+          }
+        }
+      }
+    });
+    await this.normalizeInscriptionLocations({ inscription_id: Array.from(updatedInscriptionIds) });
+    await this.refreshMaterializedView('chain_tip');
+    await this.refreshMaterializedView('inscription_count');
+    await this.refreshMaterializedView('mime_type_counts');
+    await this.refreshMaterializedView('sat_rarity_counts');
   }
 
   async getChainTipBlockHeight(): Promise<number> {
@@ -66,10 +163,10 @@ export class PgStore extends BasePgStore {
   }
 
   async getChainTipInscriptionCount(): Promise<number> {
-    const result = await this.sql<{ inscription_count: number }[]>`
-      SELECT inscription_count FROM chain_tip
+    const result = await this.sql<{ count: number }[]>`
+      SELECT count FROM inscription_count
     `;
-    return result[0].inscription_count;
+    return result[0].count;
   }
 
   async getMimeTypeInscriptionCount(mimeType?: string[]): Promise<number> {
@@ -102,189 +199,6 @@ export class PgStore extends BasePgStore {
   async getInscriptionTransfersETag(): Promise<string> {
     const result = await this.sql<{ max: number }[]>`SELECT MAX(id) FROM locations`;
     return result[0].max.toString();
-  }
-
-  async insertInscriptionGenesis(args: {
-    inscription: DbInscriptionInsert;
-    location: DbLocationInsert;
-  }): Promise<void> {
-    let inscription_id: number | undefined;
-    await this.sqlWriteTransaction(async sql => {
-      // Are we upserting?
-      const prevInscription = await sql<{ id: number }[]>`
-        SELECT id FROM inscriptions WHERE number = ${args.inscription.number}
-      `;
-      if (prevInscription.count !== 0) {
-        logger.warn(
-          {
-            block_height: args.location.block_height,
-            genesis_id: args.inscription.genesis_id,
-          },
-          `PgStore upserting inscription genesis #${args.inscription.number}`
-        );
-      } else {
-        // Is this a sequential genesis insert?
-        const maxNumber = await this.getMaxInscriptionNumber();
-        if (maxNumber && maxNumber + 1 !== args.inscription.number) {
-          logger.error(
-            {
-              block_height: args.location.block_height,
-              genesis_id: args.inscription.genesis_id,
-            },
-            `PgStore inserting out-of-order inscription genesis #${args.inscription.number}, current max is #${maxNumber}`
-          );
-        }
-      }
-
-      const inscription = await sql<{ id: number }[]>`
-        INSERT INTO inscriptions ${sql(args.inscription)}
-        ON CONFLICT ON CONSTRAINT inscriptions_number_unique DO UPDATE SET
-          genesis_id = EXCLUDED.genesis_id,
-          mime_type = EXCLUDED.mime_type,
-          content_type = EXCLUDED.content_type,
-          content_length = EXCLUDED.content_length,
-          content = EXCLUDED.content,
-          fee = EXCLUDED.fee
-        RETURNING id
-      `;
-      inscription_id = inscription[0].id;
-      const location = {
-        inscription_id,
-        block_height: args.location.block_height,
-        block_hash: args.location.block_hash,
-        tx_id: args.location.tx_id,
-        address: args.location.address,
-        output: args.location.output,
-        offset: args.location.offset,
-        value: args.location.value,
-        sat_ordinal: args.location.sat_ordinal,
-        sat_rarity: args.location.sat_rarity,
-        sat_coinbase_height: args.location.sat_coinbase_height,
-        timestamp: sql`to_timestamp(${args.location.timestamp})`,
-      };
-      await sql<DbLocation[]>`
-        INSERT INTO locations ${sql(location)}
-        ON CONFLICT ON CONSTRAINT locations_inscription_id_block_height_unique DO UPDATE SET
-          block_hash = EXCLUDED.block_hash,
-          tx_id = EXCLUDED.tx_id,
-          address = EXCLUDED.address,
-          output = EXCLUDED.output,
-          "offset" = EXCLUDED.offset,
-          value = EXCLUDED.value,
-          sat_ordinal = EXCLUDED.sat_ordinal,
-          sat_rarity = EXCLUDED.sat_rarity,
-          sat_coinbase_height = EXCLUDED.sat_coinbase_height,
-          timestamp = EXCLUDED.timestamp
-      `;
-      const json = inscriptionContentToJson(args.inscription);
-      if (json) {
-        const values = {
-          inscription_id,
-          p: json.p,
-          op: json.op,
-          content: json,
-        };
-        await sql`
-          INSERT INTO json_contents ${sql(values)}
-          ON CONFLICT ON CONSTRAINT json_contents_inscription_id_unique DO UPDATE SET
-            p = EXCLUDED.p,
-            op = EXCLUDED.op,
-            content = EXCLUDED.content
-        `;
-      }
-      await this.updateChainTipBlockHeight({ blockHeight: args.location.block_height });
-    });
-    if (inscription_id) await this.normalizeInscriptionLocations({ inscription_id });
-  }
-
-  async insertInscriptionTransfer(args: { location: DbLocationInsert }): Promise<void> {
-    let inscription_id: number | undefined;
-    await this.sqlWriteTransaction(async sql => {
-      const inscription = await sql<{ id: number }[]>`
-        SELECT id FROM inscriptions WHERE genesis_id = ${args.location.genesis_id}
-      `;
-      if (inscription.count === 0) {
-        logger.warn(
-          args.location,
-          `PgStore ignoring transfer for an inscription that does not exist`
-        );
-        return;
-      }
-      inscription_id = inscription[0].id;
-      const location = {
-        inscription_id,
-        block_height: args.location.block_height,
-        block_hash: args.location.block_hash,
-        tx_id: args.location.tx_id,
-        address: args.location.address,
-        output: args.location.output,
-        offset: args.location.offset,
-        value: args.location.value,
-        sat_ordinal: args.location.sat_ordinal,
-        sat_rarity: args.location.sat_rarity,
-        sat_coinbase_height: args.location.sat_coinbase_height,
-        timestamp: sql`to_timestamp(${args.location.timestamp})`,
-      };
-      await sql`
-        INSERT INTO locations ${sql(location)}
-        ON CONFLICT ON CONSTRAINT locations_inscription_id_block_height_unique DO UPDATE SET
-          block_hash = EXCLUDED.block_hash,
-          tx_id = EXCLUDED.tx_id,
-          address = EXCLUDED.address,
-          output = EXCLUDED.output,
-          "offset" = EXCLUDED.offset,
-          value = EXCLUDED.value,
-          sat_ordinal = EXCLUDED.sat_ordinal,
-          sat_rarity = EXCLUDED.sat_rarity,
-          sat_coinbase_height = EXCLUDED.sat_coinbase_height,
-          timestamp = EXCLUDED.timestamp
-      `;
-    });
-    if (inscription_id) await this.normalizeInscriptionLocations({ inscription_id });
-  }
-
-  async rollBackInscriptionGenesis(args: { genesis_id: string }): Promise<void> {
-    // This will cascade into dependent tables.
-    await this.sql`DELETE FROM inscriptions WHERE genesis_id = ${args.genesis_id}`;
-  }
-
-  async rollBackInscriptionTransfer(args: { genesis_id: string; output: string }): Promise<void> {
-    let inscription_id: number | undefined;
-    await this.sqlWriteTransaction(async sql => {
-      const inscription = await sql<{ id: number }[]>`
-        SELECT id FROM inscriptions WHERE genesis_id = ${args.genesis_id}
-      `;
-      if (inscription.count === 0) {
-        logger.warn(args, `PgStore ignoring rollback for a transfer that does not exist`);
-        return;
-      }
-      inscription_id = inscription[0].id;
-      await sql`
-        DELETE FROM locations
-        WHERE inscription_id = ${inscription_id} AND output = ${args.output}
-      `;
-    });
-    if (inscription_id) await this.normalizeInscriptionLocations({ inscription_id });
-  }
-
-  private async normalizeInscriptionLocations(args: { inscription_id: number }): Promise<void> {
-    await this.sql`
-      WITH i_genesis AS (
-        SELECT id FROM locations
-        WHERE inscription_id = ${args.inscription_id}
-        ORDER BY block_height ASC
-        LIMIT 1
-      ), i_current AS (
-        SELECT id FROM locations
-        WHERE inscription_id = ${args.inscription_id}
-        ORDER BY block_height DESC
-        LIMIT 1
-      )
-      UPDATE locations SET
-        current = (CASE WHEN id = (SELECT id FROM i_current) THEN TRUE ELSE FALSE END),
-        genesis = (CASE WHEN id = (SELECT id FROM i_genesis) THEN TRUE ELSE FALSE END)
-      WHERE inscription_id = ${args.inscription_id}
-    `;
   }
 
   async getInscriptionCurrentLocation(args: { output: string }): Promise<DbLocation | undefined> {
@@ -340,7 +254,8 @@ export class PgStore extends BasePgStore {
     sort?: DbInscriptionIndexOrder
   ): Promise<DbPaginatedResult<DbFullyLocatedInscriptionResult>> {
     return await this.sqlTransaction(async sql => {
-      // Do we need a filtered `COUNT(*)`?
+      // Do we need a filtered `COUNT(*)`? If so, try to use the pre-calculated counts we have in
+      // materialized views to speed up these queries.
       const countType = getIndexResultCountType(filters);
       // `ORDER BY` statement
       let orderBy = sql`gen.block_height`;
@@ -517,5 +432,196 @@ export class PgStore extends BasePgStore {
     await this.sql`REFRESH MATERIALIZED VIEW ${
       isProd ? this.sql`CONCURRENTLY` : this.sql``
     } ${this.sql(viewName)}`;
+  }
+
+  private async insertInscriptionGenesis(args: {
+    inscription: DbInscriptionInsert;
+    location: DbLocationInsert;
+  }): Promise<number | undefined> {
+    let inscription_id: number | undefined;
+    await this.sqlWriteTransaction(async sql => {
+      // Are we upserting?
+      const prevInscription = await sql<{ id: number }[]>`
+        SELECT id FROM inscriptions WHERE number = ${args.inscription.number}
+      `;
+      if (prevInscription.count !== 0) {
+        logger.warn(
+          {
+            block_height: args.location.block_height,
+            genesis_id: args.inscription.genesis_id,
+          },
+          `PgStore upserting inscription genesis #${args.inscription.number}`
+        );
+      } else {
+        // Is this a sequential genesis insert?
+        const maxNumber = await this.getMaxInscriptionNumber();
+        if (maxNumber && maxNumber + 1 !== args.inscription.number) {
+          logger.error(
+            {
+              block_height: args.location.block_height,
+              genesis_id: args.inscription.genesis_id,
+            },
+            `PgStore inserting out-of-order inscription genesis #${args.inscription.number}, current max is #${maxNumber}`
+          );
+        }
+      }
+
+      const inscription = await sql<{ id: number }[]>`
+        INSERT INTO inscriptions ${sql(args.inscription)}
+        ON CONFLICT ON CONSTRAINT inscriptions_number_unique DO UPDATE SET
+          genesis_id = EXCLUDED.genesis_id,
+          mime_type = EXCLUDED.mime_type,
+          content_type = EXCLUDED.content_type,
+          content_length = EXCLUDED.content_length,
+          content = EXCLUDED.content,
+          fee = EXCLUDED.fee
+        RETURNING id
+      `;
+      inscription_id = inscription[0].id;
+      const location = {
+        inscription_id,
+        block_height: args.location.block_height,
+        block_hash: args.location.block_hash,
+        tx_id: args.location.tx_id,
+        address: args.location.address,
+        output: args.location.output,
+        offset: args.location.offset,
+        value: args.location.value,
+        sat_ordinal: args.location.sat_ordinal,
+        sat_rarity: args.location.sat_rarity,
+        sat_coinbase_height: args.location.sat_coinbase_height,
+        timestamp: sql`to_timestamp(${args.location.timestamp})`,
+      };
+      await sql<DbLocation[]>`
+        INSERT INTO locations ${sql(location)}
+        ON CONFLICT ON CONSTRAINT locations_inscription_id_block_height_unique DO UPDATE SET
+          block_hash = EXCLUDED.block_hash,
+          tx_id = EXCLUDED.tx_id,
+          address = EXCLUDED.address,
+          output = EXCLUDED.output,
+          "offset" = EXCLUDED.offset,
+          value = EXCLUDED.value,
+          sat_ordinal = EXCLUDED.sat_ordinal,
+          sat_rarity = EXCLUDED.sat_rarity,
+          sat_coinbase_height = EXCLUDED.sat_coinbase_height,
+          timestamp = EXCLUDED.timestamp
+      `;
+      const json = inscriptionContentToJson(args.inscription);
+      if (json) {
+        const values = {
+          inscription_id,
+          p: json.p,
+          op: json.op,
+          content: json,
+        };
+        await sql`
+          INSERT INTO json_contents ${sql(values)}
+          ON CONFLICT ON CONSTRAINT json_contents_inscription_id_unique DO UPDATE SET
+            p = EXCLUDED.p,
+            op = EXCLUDED.op,
+            content = EXCLUDED.content
+        `;
+      }
+    });
+    return inscription_id;
+  }
+
+  private async insertInscriptionTransfer(args: {
+    location: DbLocationInsert;
+  }): Promise<number | undefined> {
+    let inscription_id: number | undefined;
+    await this.sqlWriteTransaction(async sql => {
+      const inscription = await sql<{ id: number }[]>`
+        SELECT id FROM inscriptions WHERE genesis_id = ${args.location.genesis_id}
+      `;
+      if (inscription.count === 0) {
+        logger.warn(
+          args.location,
+          `PgStore ignoring transfer for an inscription that does not exist`
+        );
+        return;
+      }
+      inscription_id = inscription[0].id;
+      const location = {
+        inscription_id,
+        block_height: args.location.block_height,
+        block_hash: args.location.block_hash,
+        tx_id: args.location.tx_id,
+        address: args.location.address,
+        output: args.location.output,
+        offset: args.location.offset,
+        value: args.location.value,
+        sat_ordinal: args.location.sat_ordinal,
+        sat_rarity: args.location.sat_rarity,
+        sat_coinbase_height: args.location.sat_coinbase_height,
+        timestamp: sql`to_timestamp(${args.location.timestamp})`,
+      };
+      await sql`
+        INSERT INTO locations ${sql(location)}
+        ON CONFLICT ON CONSTRAINT locations_inscription_id_block_height_unique DO UPDATE SET
+          block_hash = EXCLUDED.block_hash,
+          tx_id = EXCLUDED.tx_id,
+          address = EXCLUDED.address,
+          output = EXCLUDED.output,
+          "offset" = EXCLUDED.offset,
+          value = EXCLUDED.value,
+          sat_ordinal = EXCLUDED.sat_ordinal,
+          sat_rarity = EXCLUDED.sat_rarity,
+          sat_coinbase_height = EXCLUDED.sat_coinbase_height,
+          timestamp = EXCLUDED.timestamp
+      `;
+    });
+    return inscription_id;
+  }
+
+  private async rollBackInscriptionGenesis(args: { genesis_id: string }): Promise<void> {
+    // This will cascade into dependent tables.
+    await this.sql`DELETE FROM inscriptions WHERE genesis_id = ${args.genesis_id}`;
+  }
+
+  private async rollBackInscriptionTransfer(args: {
+    genesis_id: string;
+    output: string;
+  }): Promise<number | undefined> {
+    let inscription_id: number | undefined;
+    await this.sqlWriteTransaction(async sql => {
+      const inscription = await sql<{ id: number }[]>`
+        SELECT id FROM inscriptions WHERE genesis_id = ${args.genesis_id}
+      `;
+      if (inscription.count === 0) {
+        logger.warn(args, `PgStore ignoring rollback for a transfer that does not exist`);
+        return;
+      }
+      inscription_id = inscription[0].id;
+      await sql`
+        DELETE FROM locations
+        WHERE inscription_id = ${inscription_id} AND output = ${args.output}
+      `;
+    });
+    return inscription_id;
+  }
+
+  private async normalizeInscriptionLocations(args: { inscription_id: number[] }): Promise<void> {
+    await this.sqlWriteTransaction(async sql => {
+      for (const id of args.inscription_id) {
+        await sql`
+          WITH i_genesis AS (
+            SELECT id FROM locations
+            WHERE inscription_id = ${id}
+            ORDER BY block_height ASC
+            LIMIT 1
+          ), i_current AS (
+            SELECT id FROM locations
+            WHERE inscription_id = ${id}
+            ORDER BY block_height DESC
+            LIMIT 1
+          )
+          UPDATE locations SET
+            current = (CASE WHEN id = (SELECT id FROM i_current) THEN TRUE ELSE FALSE END),
+            genesis = (CASE WHEN id = (SELECT id FROM i_genesis) THEN TRUE ELSE FALSE END)
+          WHERE inscription_id = ${id}
+        `;
+      }
+    });
   }
 }
